+++ prover/src/tables/trace_builder.rs
@@ -3799,74 +3799,4 @@ impl Traces {
         )
     }
 
-    /// Generates all traces with a trimmed bitwise table (TEST ONLY).
-    ///
-    /// # WARNING: UNSOUND FOR PRODUCTION
-    ///
-    /// This function generates the full 2^20 row bitwise table, updates multiplicities,
-    /// then removes rows where all multiplicity columns are zero. This is **unsound**
-    /// because:
-    ///
-    /// 1. The bitwise table is NOT preprocessed - the verifier checks the prover's
-    ///    commitment instead of a hardcoded trusted commitment
-    /// 2. A malicious prover could provide incorrect bitwise results and the
-    ///    verifier would accept them (e.g., claim 5 AND 3 = 7)
-    /// 3. The table structure differs from production (row indices don't match)
-    ///
-    /// This is acceptable for tests because we're testing:
-    /// - Bus interaction balancing (sends = receives)
-    /// - Constraint satisfaction
-    /// - LogUp protocol correctness
-    ///
-    /// The full preprocessed bitwise verification is tested separately in the
-    /// comprehensive `test_prove_elfs_all_instructions_64_full` test.
-    #[cfg(test)]
-    pub fn from_logs_trimmed(
-        logs: &[Log],
-        instructions: U64HashMap<Instruction>,
-        max_rows: &super::MaxRowsConfig,
-    ) -> Result<Self, Error> {
-        // Generate full traces (including full 2^20 bitwise table with multiplicities)
-        let mut traces = Self::from_logs(logs, instructions, max_rows)?;
-
-        // Trim the bitwise table to only rows with non-zero multiplicities
-        traces.bitwise = bitwise::trim_zero_rows(traces.bitwise);
-
-        Ok(traces)
-    }
-
-    /// Generates all traces with a minimal bitwise table (TEST ONLY).
-    ///
-    /// Alias for `from_logs_trimmed` for backwards compatibility.
-    #[cfg(test)]
-    pub fn from_logs_minimal(
-        logs: &[Log],
-        instructions: U64HashMap<Instruction>,
-        max_rows: &super::MaxRowsConfig,
-    ) -> Result<Self, Error> {
-        Self::from_logs_trimmed(logs, instructions, max_rows)
-    }
-
-    /// Like [`from_elf_and_logs`] but trims the bitwise table (TEST ONLY).
-    ///
-    /// Produces PAGE and REGISTER tables (requires ELF) while keeping the
-    /// bitwise table small. Same unsoundness caveats as [`from_logs_trimmed`].
-    #[cfg(test)]
-    pub fn from_elf_and_logs_minimal(
-        elf: &Elf,
-        logs: &[Log],
-        max_rows: &super::MaxRowsConfig,
-        private_input: &[u8],
-    ) -> Result<Self, Error> {
-        let mut traces = Self::from_elf_and_logs(
-            elf,
-            logs,
-            max_rows,
-            private_input,
-            #[cfg(feature = "disk-spill")]
-            StorageMode::Ram,
-        )?;
-        traces.bitwise = bitwise::trim_zero_rows(traces.bitwise);
-        Ok(traces)
-    }
 }
